requirements.txt
finnhub-python==2.4.19
pandas
requests
jinja2